Make CopyToClipboard generic
