+++ Flow.Launcher/MainWindow.xaml.cs
@@ -63,8 +63,7 @@ namespace Flow.Launcher
             if (QueryTextBox.SelectionLength == 0 && result != null)
             {
                 string copyText = result.CopyText;
-                _viewModel.ResultCopy(copyText);
-
+                App.API.CopyToClipboard(copyText, directCopy: true);
             }
             else if (!string.IsNullOrEmpty(QueryTextBox.Text))
             {
