+++ Flow.Launcher/ViewModel/MainViewModel.cs
@@ -1103,43 +1103,6 @@ namespace Flow.Launcher.ViewModel
             Results.AddResults(resultsForUpdates, token);
         }
 
-        /// <summary>
-        /// Copies the specified file or folder path to the clipboard, or the specified text if it is not a valid file or folder path.
-        /// Shows a message indicating whether the operation was completed successfully.
-        /// </summary>
-        /// <param name="stringToCopy">The file or folder path, or text to copy to the clipboard.</param>
-        /// <returns>Nothing.</returns>
-        public void ResultCopy(string stringToCopy)
-        {
-            if (string.IsNullOrEmpty(stringToCopy))
-                return;
-
-            var isFile = File.Exists(stringToCopy);
-            var isFolder = isFile ? false : Directory.Exists(stringToCopy); // No need to eval directory exists if determined that file exists 
-            if (isFile || isFolder)
-            {
-                var paths = new StringCollection
-                {
-                    stringToCopy
-                };
-
-                Clipboard.SetFileDropList(paths);
-                
-                App.API.ShowMsg(
-                    $"{App.API.GetTranslation("copy")} {(isFile ? App.API.GetTranslation("fileTitle") : App.API.GetTranslation("folderTitle"))}",
-                    App.API.GetTranslation("completedSuccessfully"));
-            }
-            else
-            {
-                Clipboard.SetDataObject(stringToCopy);
-                
-                App.API.ShowMsg(
-                    $"{App.API.GetTranslation("copy")} {App.API.GetTranslation("textTitle")}",
-                    App.API.GetTranslation("completedSuccessfully"));
-            }
-            return;
-        }
-
         #endregion
     }
 }
